import basic mp3, read mp3 tags

+++ src/main/java/ch/fhnw/webec/entity/Book.java
@@ -13,7 +13,7 @@ public class Book {
     @ManyToOne(cascade = CascadeType.ALL)
     @JoinColumn(name = "author_id")
     private Author author;
-    private Integer length;
+    private Long length;
     private String filename;
     private String title;
 
@@ -25,11 +25,11 @@ public class Book {
         this.author = author;
     }
 
-    public Integer getLength() {
+    public Long getLength() {
         return length;
     }
 
-    public void setLength(Integer length) {
+    public void setLength(Long length) {
         this.length = length;
     }
